+++ fj-doc-maven-plugin/src/main/java/org/fugerit/java/doc/project/facade/BasicVenusFacade.java
@@ -113,34 +113,39 @@ public class BasicVenusFacade {
 
     private static void addPlugin( VenusContext context, Model model ) throws IOException {
         if ( context.isAddVerifyPlugin() ) {
-            Build build = model.getBuild();
-            if ( build == null ) {
-                build = new Build();
-                model.setBuild( build );
-            }
-            List<Plugin> plugins = model.getBuild().getPlugins();
-            Plugin plugin = new Plugin();
-            plugin.setGroupId( GROUP_ID );
-            plugin.setArtifactId( "fj-doc-maven-plugin" );
-            plugin.setVersion( "${"+KEY_VERSION+"}" );
-            PluginExecution execution = new PluginExecution();
-            execution.setId( "freemarker-verify" );
-            execution.setPhase( "compile" );
-            execution.addGoal( "verify" );
-            plugin.getExecutions().add( execution );
-            String xml = "<configuration>\n" +
-                    "      <templateBasePath>${project.basedir}/src/main/resources/"+context.getArtificatIdForFolder()+"/template</templateBasePath>\n" +
-                    "      <generateReport>true</generateReport>\n" +
-                    "      <failOnErrors>true</failOnErrors>\n" +
-                    "      <reportOutputFolder>${project.build.directory}/freemarker-syntax-verify-report</reportOutputFolder>\n" +
-                    "    </configuration>";
-            HelperIOException.apply( () -> {
-                try ( StringReader sr = new StringReader( xml ) ) {
-                    Xpp3Dom dom = Xpp3DomBuilder.build( sr );
-                    plugin.setConfiguration( dom );
+            if ( context.isVerifyPluginNotAvailable() ) {
+                log.info( "addVerifyPlugin skipped, version {} has been selected, minimum required version is : {}", context.getVersion(), VenusContext.VERSION_NA_VERIFY_PLUGIN );
+            } else {
+                log.info( "addVerifyPlugin, version {} has been selected, minimum required version is : {}", context.getVersion(), VenusContext.VERSION_NA_VERIFY_PLUGIN );
+                Build build = model.getBuild();
+                if ( build == null ) {
+                    build = new Build();
+                    model.setBuild( build );
                 }
-            });
-            plugins.add( plugin );
+                List<Plugin> plugins = model.getBuild().getPlugins();
+                Plugin plugin = new Plugin();
+                plugin.setGroupId( GROUP_ID );
+                plugin.setArtifactId( "fj-doc-maven-plugin" );
+                plugin.setVersion( "${"+KEY_VERSION+"}" );
+                PluginExecution execution = new PluginExecution();
+                execution.setId( "freemarker-verify" );
+                execution.setPhase( "compile" );
+                execution.addGoal( "verify" );
+                plugin.getExecutions().add( execution );
+                String xml = "<configuration>\n" +
+                        "      <templateBasePath>${project.basedir}/src/main/resources/"+context.getArtificatIdForFolder()+"/template</templateBasePath>\n" +
+                        "      <generateReport>true</generateReport>\n" +
+                        "      <failOnErrors>true</failOnErrors>\n" +
+                        "      <reportOutputFolder>${project.build.directory}/freemarker-syntax-verify-report</reportOutputFolder>\n" +
+                        "    </configuration>";
+                HelperIOException.apply( () -> {
+                    try ( StringReader sr = new StringReader( xml ) ) {
+                        Xpp3Dom dom = Xpp3DomBuilder.build( sr );
+                        plugin.setConfiguration( dom );
+                    }
+                });
+                plugins.add( plugin );
+            }
         }
     }